feat: 등록된 이미지 길게 눌러 삭제하는 기능 구현

+++ AR-GPT-ImageDetection/AR-GPT-ImageDetection/CoreData/Utility/Implementation/MarkerImageManager.swift
@@ -76,8 +76,10 @@ extension MarkerImageManager: MarkerImageManageable {
     }
     
     // 마커 삭제
-    func delete(item: MarkerImageMO) {
-        persistentContainer.viewContext.delete(item)
+    func delete(by id: UUID) {
+        if let item = fetch(by: id) {
+            persistentContainer.viewContext.delete(item)            
+        }
         save()
     }
 }
